Move all lib files to dedicated subroutes
