+++ lib/field/field.tsx
@@ -12,8 +12,8 @@ import {
   useState,
 } from "react";
 import { FieldInstanceProps, FieldInstance } from "./types";
-import { FormContext } from "./form-context";
-import { getValidationError, stringToPath, validate } from "./utils";
+import { FormContext } from "../form/context";
+import { getValidationError, stringToPath, validate } from "../utils";
 
 export interface FieldRenderProps<T = any> extends FieldInstanceProps<T> {
   children: (props: FieldInstance<T>) => JSX.Element;
