+++ lib/field/types.ts
@@ -1,5 +1,5 @@
 import { ZodTypeAny } from "zod";
-import { FormContext } from "./form-context";
+import { FormContext } from "../form/context";
 
 type ValidationFunction<T> =
   | ZodTypeAny
