+++ lib/utils/path.ts
@@ -1,26 +1,3 @@
-import { ZodError, ZodTypeAny } from "zod";
-import { FormContext, initialContext } from "./form-context";
-
-export function validate<T>(
-  val: T,
-  form: FormContext<T>,
-  validator: ZodTypeAny | ((val: T, form: FormContext<T>) => Promise<boolean>)
-) {
-  if (validator instanceof Function) {
-    return validator(val, form);
-  } else {
-    return validator.parseAsync(val);
-  }
-}
-
-export function getValidationError(error: ZodError | string) {
-  if (error instanceof ZodError) {
-    return error.errors.map((error) => error.message);
-  } else {
-    return [error];
-  }
-}
-
 /** Used to match property names within property paths. */
 const rePropName =
   /[^.[\]]+|\[(?:(-?\d+(?:\.\d+)?)|(["'])((?:(?!\2)[^\\]|\\.)*?)\2)\]|(?=(?:\.|\[\])(?:\.|\[\]|$))/g;
